Prove that version shows up in CMakeLists
There doesn't seem to be a great way to try running cmake
and asserting that it fails in the 'installation' group. It
seems likely that there could easily be false positive
failures, and CMake is really slow to configure compared to
the other tests and I want to avoid adding more configures.