Improve CLANG_CHECK expectations for running unit tests
Currently it's expected that the unit tests have a CLANG_CHECK
environment variable set to a clang-check executable. This was not
obvious in the test and so the assertion was made more clear.

It also seems like 99% of the time after getting things set up on a
machine, this doesn't need to be a manual process during development.
Added a quick implementation for automatically setting on macOS. Windows
and Linux remain to be done.