Closes #34, remove TODO for -lc++ in pkgconfig
This is somewhat dicey, but it seems to be the best decision
to not force pkg-config consumers to use a specific c++
library implementation. This is only a problem when a
consuming body of _C_ code _statically_ links to a library
implemented with the standard C++ library. At that point,
_what is the point of being plain C?_. The runtime is already
going to link to C++. If the developer prefers to write in
plain C syntax, that is perfectly acceptable, but the dev can
just use C++ tooling to link the standard library.