fix: Make MSVC use /MDd by default
The implementation is to always include a `/MDd` flag. This could
be somewhat controversial. It seems like this is the default in
Visual Studio and in CMake from what I'm reading and experiencing.
It also seems like the obvious choice to not statically link the
CRT, though I could be naive to Windows systems and maybe that's
wrong. In any case, it's overridable by including whichever CRT
linking strategy is desired in the `esmakefile-cmake.config.json`
file with `cflags` and `cxxflags`. Users who need a different
strategy can use these config files to use the less common options.

As proof of overriding, `cl.exe` will explicitly call out an
override, proving that it specifically handles it:

```
cl : Command line warning D9025 : overriding '/MDd' with '/MT'
```